add basic implementation

+++ api/src/main/java/me/dirantos/moneymaker/api/Interest.java
@@ -2,6 +2,6 @@ package me.dirantos.moneymaker.api;
 
 public interface Interest extends Transaction {
 
-    float getInterestRate();
+    double getInterestRate();
 
 }
